// src/types/alarmtypes.ts
export type alarmtypedetailtype = {
    id: string;
    name: string;
    comment: string;
    owner_id: string;
    owner_username: string;
    time_created: string;
    time_modified: string;
    alarm_definition?: {
      low_severity?: {
        conditions:
          | [
              {
                index: number;
                parameter: string;
                operator: string;
                coef: number;
                value: string;
                logical_operator: string;
              },
            ]
          | [];
        fault: string;
      };
      medium_severity?: {
        conditions:
          | [
              {
                index: number;
                parameter: string;
                operator: string;
                coef: number;
                value: string;
                logical_operator: string;
              },
            ]
          | [];
        fault: string;
      };
      high_severity?: {
        conditions:
          | [
              {
                index: number;
                parameter: string;
                operator: string;
                coef: number;
                value: string;
                logical_operator: string;
              },
            ]
          | [];
        fault: string;
      };
    };
    alarm_content: {
      primary_source: string;
      secondary_source: string;
      alarm_details: {
        date_and_time: string[];
        network: string[];
        rtu: string[];
        optical_route: string[];
        test_setup: string[];
        test_result: string[];
      };
    };
    alert_sending: {
      about: string;
      user: [string] | [];
    };
    automatic_events: {
      escalate_alarm: {
        severity_at_least: string;
        escalate_pending_after: {
          days: number;
          hours: number;
          minutes: number;
        };
        escalate_acknowledged_after: {
          days: number;
          hours: number;
          minutes: number;
        };
      };
      timeout_alarm: {
        timeout_pending_after: {
          days: number;
          hours: number;
          minutes: number;
        };
        timeout_acknowledged_after: {
          days: number;
          hours: number;
          minutes: number;
        };
      };
      delete_alarm: {
        delete_resolved_after: {
          days: number;
          hours: number;
          minutes: number;
        };
        delete_in_progress_after: {
          days: number;
          hours: number;
          minutes: number;
        };
        delete_timeout_after: {
          days: number;
          hours: number;
          minutes: number;
        };
      };
    };
    alarm_networks: {
      network_id_list: {id: string; name: string}[];
    };
  };
  
  export type alarmtypelist = {
    id: string;
    name: string;
  }[];